codigos/ocr/requirements.txt
streamlit>=1.28.0
pdf2image>=1.16.3
Pillow>=10.0.0
boto3>=1.28.0
mistralai>=1.8.1
python-docx>=1.1.2
pydantic>=2.0.0
python-dotenv>=1.0.0
repenseai>=4.0.13